use uint32 for the metric value

+++ modules/blockbuilder/live_traces_iter_test.go
@@ -43,5 +43,5 @@ func TestLiveTracesIter_DedupSpans(t *testing.T) {
 		}
 	}
 	require.Equal(t, spanCount, total)
-	require.Equal(t, spanCount, iter.DedupedSpans())
+	require.Equal(t, uint32(spanCount), iter.DedupedSpans())
 }
